[#21] Small refactoring for web-service: step 6.5 (services refactoring)
