+++ src/podcast/ViewModels/LoginViewModel.swift
@@ -6,19 +6,17 @@ class LoginViewModel: ObservableObject {
     var email: String = ""
     var password: String = ""
     var token: String = ""
+
     @Published var isAuthenticated: Bool = false
     @Published var notifyUserIsAuthenticated: Bool = false
 
     
     func login(){
         
-//        let defaults = UserDefaults.standard
-        
-        WebService().login(email: self.email, password: self.password){ result in
+        AuthService().login(email: self.email, password: self.password){ result in
             switch result{
                 case .success(let token):
                     print("Access token \(token)")
-//                    defaults.setValue(token, forKey: "accessToken")
                     
                     let keychain = Keychain(service: "com.podcast")
                     do {
